refactor(ocr): extract provider abstraction for multi-model support (#50)
Split OCR into provider interface + SiliconFlow implementation + factory.
env vars SILICONFLOW_MODEL / SILICONFLOW_API_URL allow model/endpoint
override without code changes. OCR_PROVIDER env selects provider.

Backwards compatible: old import path @/lib/ocr/scoreboard still works.

Constraint: no behavior change, extract + validate flow unchanged
Confidence: high
Scope-risk: narrow
